Solve the warning message and pass the GPIO_IN_OUT_test.py testing script.

+++ source/py_gpio.c
@@ -251,7 +251,7 @@ static PyObject *py_setup_channel(PyObject *self, PyObject *args, PyObject *kwar
 		if(setupflag==0)
 		{
 			PyErr_SetString(PyExc_RuntimeError, "This gpio is set to other function!");
-			return NULL;
+			return (int)NULL;
 		}
 		gpio_direction[gpio] = direction;
 		return 1;
